+++ src/containers/DefaultHeaderDropdownAccnt.vue
@@ -56,37 +56,6 @@
         itemsCount: 42,
         error: false
       }
-    },
-    methods: {
-      logout() {
-        console.log("Start Logout. Token: " + localStorage.token);
-        let headers = {
-          headers: {
-            'Accept': 'application/json',
-            'Authorization': 'Bearer ' + localStorage.token
-          }
-        };
-
-        this.$http.post('/auth/logout', [], headers)
-            .then(request => this.logoutSuccessful(request))
-            .catch(request => this.logoutFailed(request));
-      },
-      logoutSuccessful(req) {
-        // console.log(req);
-        delete localStorage.token;
-        this.allToFalse();
-        this.userStoreConfig("");
-        this.error = false;
-        console.log("Logout!");
-        this.$router.replace('/auth/sign-in');
-      },
-      logoutFailed(req) {
-        console.log(req);
-        this.error = true;
-        this.message = 'Login failed!';
-        // console.log("Logout Failed. Token: " + localStorage);
-        // console.log("store.state.user: " + store.state.user);
-      }
     }
   }
 </script>